requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
vosk==0.3.45
openai-whisper==20231117
pyttsx3==2.90
python-multipart==0.0.6
websockets==12.0
aiohttp==3.9.1
redis==5.0.1
python-dotenv==1.0.0
sounddevice==0.4.6
numpy==1.24.3